Use tolerate-unready-endpoints for headless svc
This annotation allows the individual Pods to get DNS names ahead of
passing the readiness check. The DNS names are essential for the cluster
members to discover and connect with one another. We have plans to
enhance the readiness probe to require that a Pod has connected to the
cluster before passing; without this annotation we would deadlock.